requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
supabase==2.3.0
groq==0.4.2
python-dotenv==1.0.0
httpx>=0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
